Object Structure Rework Done
The rolling stock objects are now completely restructured: Many arrays were replaced with objects, more convenient naming was being introduced and the hierarchy now makes so much more sense. A big step to make the base future proof. Users are now able to provide a rolling stock name in 'game.js' and the algorithm will automatically capture the configuration for this unit in 'units.js'.